requirements.txt
grpcio==1.33.2
grpcio-tools==1.33.2
protobuf==3.14.0
python-dotenv==0.15.0
six==1.15.0
